refactor: add all routes logic in controllers * add external js file in public directory to handel edit request
