+++ blog.json
@@ -1,1 +1,1 @@
-{"blogs":[{"title":"JavaScript","author":"Sahil","content":"Everything is JavaScript, well well...\nJavaScript is Everything","id":1685339035142},{"title":"nodejs","author":"Sahil","content":"Using argv you can always access arguments that are passed in a command line. argv is an array which has Node itself as the first element and the absolute path of the file as the second element. From the third element onwards it can have as many arguments as you want.\r\n\r\nTry the below program to get more insight into how you can use these various properties and functions.","id":1685354322658},{"title":"Angular","author":"Bhagvanji","content":"Angular is a TypeScript-based, free and open-source web application framework led by the Angular Team at Google and by a community of individuals and corporations. Angular is a complete rewrite from the same team that built AngularJS. Wikipedia","id":1685354370218},{"title":"python","author":"Jay","content":"fsfsdfsffsfs fw  eqr","id":1685354461119}]}
+{"blogs":[{"title":"Js","author":"Sahil","content":"Js is every thing","id":1685339035142},{"title":"nodejs","author":"Sahil","content":"Using argv you can always access arguments that are passed in a command line. argv is an array which has Node itself as the first element and the absolute path of the file as the second element. From the third element onwards it can have as many arguments as you want.\r\n\r\nTry the below program to get more insight into how you can use these various properties and functions.","id":1685354322658},{"title":"Angular","author":"Bhagvanji","content":"Angular is a TypeScript-based, free and open-source web application framework led by the Angular Team at Google and by a community of individuals and corporations. Angular is a complete rewrite from the same team that built AngularJS. Wikipedia","id":1685354370218},{"title":"python","author":"Jay","content":"fsfsdfsffsfs fw  eqr","id":1685354461119}]}
